Move Text to its own drawables dir
